refactor: 폴더 구분하기
