+++ src/main/kotlin/lotto/controller/validation/Constants.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.controller.validation
 
 const val LOTTO_AMOUNT = 1000
