+++ src/main/kotlin/lotto/controller/validation/Enum.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.controller.validation
 
 enum class RANK {
     FIRST, SECOND, THIRD, FOURTH, FIFTH, OTHERS
