+++ src/main/kotlin/lotto/controller/validation/ErrorMessage.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.controller.validation
 
 const val NOT_NUMBER = "[ERROR] 입력이 숫자가 아닙니다"
 const val DUPLICATION_NUMBER = "[ERROR] 로또 번호가 중복입니다"
