+++ src/main/kotlin/lotto/controller/validation/Maps.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.controller.validation
 
 val MATCH = mapOf(RANK.FIRST to 6, RANK.SECOND to 5, RANK.THIRD to 5, RANK.FOURTH to 4, RANK.FIFTH to 3)
 val PRIZE = mapOf(RANK.FIRST to 2000000000, RANK.SECOND to 30000000,  RANK.THIRD to 1500000, RANK.FOURTH to 50000, RANK.FIFTH to 5000, RANK.OTHERS to 0)
