+++ src/main/kotlin/lotto/view/Input.kt
@@ -1,6 +1,8 @@
-package lotto
+package lotto.view
 
 import camp.nextstep.edu.missionutils.Console
+import lotto.controller.validation.NOT_NUMBER
+import lotto.controller.validation.isRangeLottoNumber
 
 class Input {
     private fun input(): String = Console.readLine()
@@ -66,7 +68,7 @@ class Input {
     }
 
     companion object {
-        private const val REQUEST_AMOUNT_MESSAGE = "구입금액을 입력해 주세요."
+        private const val REQUEST_AMOUNT_MESSAGE = "구입금액을 입력해 주세요. 예) 1000"
         private const val REQUEST_NUMBERS_MESSAGE = "당첨 번호를 입력해 주세요."
         private const val REQUEST_BONUS_NUMBERS_MESSAGE = "보너스 번호를 입력해 주세요."
     }
